Nartiãng and Nonglang Datasets
Nonglang and Nartiang are the areas where the data was collected. The main difference between the Nonglang1000 and Nonglang250, and Nartiang1000 and Nartiang250 is in the landscape data in the last few columns. One dataset reports the land cover within a 1000m radius, and the other reports the land cover within a 250m radius.